Feature/adding alert dao (#25)
* Add AlertRepository and AlertDao

* Implement AlertDAO and update AlertRepository

+++ src/main/java/com/grad/sam/repository/AlertRepository.java
@@ -1,10 +1,13 @@
 package com.grad.sam.repository;
 
+import com.grad.sam.dao.AlertDao;
+import org.springframework.beans.factory.annotation.Autowired;
 import com.grad.sam.model.Alert;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Repository
@@ -14,7 +17,7 @@ public interface AlertRepository extends JpaRepository<Alert, Integer> {
 
     List<Alert> findByStatus(String status);
 
-    List<Alert> findByAccount_AccountId(Integer accountId);
+    List<Alert> findByAccount_AccountId(Long accountId);
 
     List<Alert> findByAssignedTo(String assignedTo);
